Policy Authorization in Controller.

+++ app/Http/Controllers/JobController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
-use App\Models\Job;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     // @desc Show all job listings 
     // @route GET /jobs
     public function index(): View
@@ -79,8 +82,11 @@ class JobController extends Controller
 
     // @desc Show edit job form
     // @route GET /jobs/{$id}/edit
-    public function edit(Job $job): stirng
+    public function edit(Job $job): View
     {
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -88,6 +94,9 @@ class JobController extends Controller
     // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job): string
     {   
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         // dd($request->file('company_logo'));
         // php artisan storage:link
         $validatedData = $request->validate([
@@ -136,6 +145,9 @@ class JobController extends Controller
     // @route DELETE /jobs/{$id}
     public function destroy(Job $job): RedirectResponse
     {
+        // Check if user is authorized
+        $this->authorize('delete', $job);
+
         // If logo, then delete it
         if($job->company_logo) {
             Storage::delete('public/logos/' > $job->company_logo);
